Cleaned multiple files by moving a lot of the structs around
