+++ src/main.rs
@@ -1,11 +1,12 @@
-use leek_ast::parser::Parser;
-use leek_ast::serializer::Writer;
-use leek_ast::visitor::{Visitor, VisitorMut};
-use leek_ast::{Identifier, WithTrivia};
+use leek_ast::{
+    ast::{terminal::identifier::Identifier, trivia::with_trivia::WithTrivia},
+    parser::Parser,
+    visitor::{writer::Writer, Visitor},
+};
 
 fn main() {
     let test_str = "   /* some test */   abc def /* gh */ xyx";
-    let mut out = <Vec<WithTrivia<Identifier>> as Parser<&str>>::parse(test_str)
+    let out = <Vec<WithTrivia<Identifier>> as Parser<&str>>::parse(test_str)
         .unwrap()
         .1;
